Factor out error and option and utility
